urltestharness.html

<!doctype html>
<base id=base>
<meta charset=utf-8>
<link rel=stylesheet href=testharness.css>
<script src=testharness.js></script>
<script src=testharnessreport.js></script>
<script src=urltestparser.js></script>
<script src=urltestloader.js></script>
<div id=log></div>
<script>
var setup = async_test("Loading data…")
setup.step(setup.step_func(loadTests(function(tests) {
  runURLTests(tests)
  setup.done()
})))

function setBase(base) {
  document.getElementById("base").href = base
}

function bURL(url, base) {
  base = base || "about:blank"
  setBase(base)
  var a = document.createElement("a")
  a.setAttribute("href", url)
  return a
}

function runURLTests(raw) {
  var urltests = URLTestParser(raw)
  for(var i = 0, l = urltests.length; i < l; i++) {
    var urltest = urltests[i]
    test(function() {
      var url = bURL(urltest.input, urltest.base)

      assert_equals(url.protocol, urltest.output.protocol, "scheme")
      assert_equals(url.hostname, urltest.output.host, "host")
      assert_equals(url.port, urltest.output.port, "port")
      assert_equals(url.pathname, urltest.output.path, "path")
      assert_equals(url.search, urltest.output.search, "search")
      assert_equals(url.hash, urltest.output.hash, "hash")
      assert_equals(url.href, urltest.output.href, "href")
    }, "Parsing: <" + urltest.input + "> against <" + urltest.base + ">")
  }
}
</script>
